renamed /context/ to /contexts/
